E2E tests: move the hello gutenberg cypress test
